MRG: Merging rev 1790 into caracal branch.

+++ ocelot/cuda/test/functions/indirectCall.cu
@@ -69,7 +69,7 @@ int main(int argc, char *arg[]) {
 	
 	result = cudaMalloc((void **)&A_gpu, bytes);
 	if (result != cudaSuccess) {
-		printf("cudaMalloc() - failed to allocate %d on the device \n", bytes);
+		printf("cudaMalloc() - failed to allocate %d on the device \n", (int)bytes);
 		return 2;
 	}
 	
@@ -80,7 +80,7 @@ int main(int argc, char *arg[]) {
 	
 	result = cudaMemcpy(A_gpu, A_cpu, bytes, cudaMemcpyHostToDevice);
 	if (result != cudaSuccess) {
-		printf("cudaMemcpy() - failed to copy %d bytes TO the device\n", bytes);
+		printf("cudaMemcpy() - failed to copy %d bytes TO the device\n", (int)bytes);
 		return 2;
 	}
 	
@@ -97,7 +97,7 @@ int main(int argc, char *arg[]) {
 	
 	result = cudaMemcpy(A_cpu, A_gpu, bytes, cudaMemcpyDeviceToHost);
 	if (result != cudaSuccess) {
-		printf("cudaMemcpy() - failed to copy %d bytes FROM the device\n", bytes);
+		printf("cudaMemcpy() - failed to copy %d bytes FROM the device\n", (int)bytes);
 		return 2;
 	}
